# python_server/docs/api.yaml
openapi: 3.0.1
info:
  title: Voteapp Python Server
  contact:
    email: dariocastellano18@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: http://localhost:8090/
tags:
- name: voting
  description: Everything about voting management
- name: algorand
  description: SDK operation toward algorand blockchain
paths:
  /voting:
    get:
      tags:
      - voting
      summary: Retrieve all Voting
      description: Retrieve all voting
      operationId: getVotings
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voting'
        400:
          description: Unknown error
          content: {}
    post:
      tags:
      - voting
      summary: Add a new voting to the db
      operationId: addVoting
      requestBody:
        description: Voting object that needs to be added to the db
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/RequestBody'
        required: true
      responses:
        405:
          description: Invalid input
          content: {}
      x-codegen-request-body-name: body
  /voting/{votingId}:
    get:
      tags:
      - voting
      summary: Find voting by ID
      description: Returns a single voting
      operationId: getVotingById
      parameters:
      - name: votingId
        in: path
        description: ID of voting to return
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voting'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: voting not found
          content: {}
    delete:
      tags:
      - voting
      summary: Deletes a voting
      operationId: deleteVoting
      parameters:
      - name: votingId
        in: path
        description: Voting id to delete
        required: true
        schema:
          type: integer
          format: int64
      responses:
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: voting not found
          content: {}
components:
  schemas:
    Voting:
      type: object
      properties:
        id:
          type: integer
          format: int64
        algorandAppId:
          type: string
        assetId:
          type: string
        description:
          type: string
        title:
          type: string
        options:
          type: array
          items:
            type: string
        creator:
          type: string
    RequestBody:
      type: object
      properties:
        assetId:
          type: string
        description:
          type: string
        title:
          type: string
        options:
          type: array
          items:
            type: string
        passphrase:
          type: string
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          type: object
          properties: {}
        message:
          type: string